JS-Performance: Dragenter und Dragleave Event-Listener entfernt JS-Performance: Query von .data-items in Bereich #SubMenu gesetzt Bugfix: Fehler beim Refreshen des Plans, wenn keine gecachte Datei existiert Bugfix: Standardpläne anlegen -> Checkboxen werden nicht richtig zurückgesetzt nach dem Anlegen Build

+++ rest/Refresh/Plan.php
@@ -6,7 +6,10 @@
  */
 
 include_once(dirname(dirname(__DIR__)) . '/config.php');
-unlink(BASE . '_cache/Plan.php');
+
+if (file_exists(BASE . '_cache/Plan.php')) {
+    unlink(BASE . '_cache/Plan.php');
+}
 
 ob_start('minifier');
 include_once(BASE . '/core/Plan.php');
